+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+Fr 18-03-2016
+Haskier a-0.3
+- Full HSF support
+- Added user-interaction features and human-like display
+- Auto-completion now supports files
+- Improved server's search engine (glob), like looking only for files or folders
+
+*** === Misc === ***
+- Intelligent auto-completion : does not interact in questions
+- Added Server.normalize()
+
+*** === Bug fixes === ***
+- Fixed bugs with JS-reserved names in servers' filesystems
+- Fixed bug with Server.fileExists() which not detects a file if it's content is an empty string ""
+
 Th 17-03-2016
 Haskier a-0.2
 - New commands : set, force-save
@@ -6,6 +21,7 @@ Haskier a-0.2
 - Created server feature, including import/export
 - Support of hidden files
 - HSF (Haskier Scenario File) are now supported [Math.js 1.7.1 parser]
+
 *** === Bug fixes === ***
 - Server was returning bad file length for plain text